requirements.txt
aws-cdk-lib>=2.114.1
constructs>=10.0.0,<11.0.0
cdk-nag==2.28.32
aws_cdk.aws_glue_alpha==2.131.0a0
faker==24.0.0
boto3==1.34.56
requests==2.31.0
